requirements.txt
flask[async]
openai
python-dotenv
pydub
sounddevice
scipy
streamlit
soundfile
quart